// src/app/admin-view/models/partyRepresentative.ts
export class PartyRepresentative {
    ID: string;
    RepMobileNo: string;
    Rep_Name: string;
    PartyId: string;
    RegistrationDate: string;
    Symbols: string;
    RoleId: number;
    DistrictId: string;
    LocalBodyId: string;
    IsActive: string;
    PartyName: string;
    MobileNo: string;
    TokenId: string;
}